requirements.txt
absl-py==2.1.0
antlr4-python3-runtime==4.9.3
asttokens==2.4.1
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
chex==0.1.85
cloudpickle==3.0.0
contextlib2==21.6.0
contourpy==1.2.0
cycler==0.12.1
decorator==4.4.2
distrax==0.1.5
dm-tree==0.1.8
etils==1.7.0
executing==2.0.1
Farama-Notifications==0.0.4
filelock==3.13.1
flax==0.7.4
fonttools==4.49.0
fsspec==2024.2.0
ftfy==6.1.3
gast==0.5.4
gdown==5.1.0
gitdb==4.0.11
GitPython==3.1.42
glfw==2.7.0
gymnasium==0.29.1
gymnasium-robotics==1.2.4
huggingface-hub==0.21.4
hydra-core==1.3.2
idna==3.6
imageio==2.34.0
imageio-ffmpeg==0.4.9
importlib_resources==6.1.3
ipython==8.22.2
jax==0.4.16
jaxlib @ https://storage.googleapis.com/jax-releases/cuda12/jaxlib-0.4.16+cuda12.cudnn89-cp311-cp311-manylinux2014_x86_64.whl#sha256=065f61497f1c2d75ff53e05a6d0d92060b9f4ac487ee16211a46121c6c245767
jedi==0.19.1
Jinja2==3.1.3
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline==0.1.6
mdurl==0.1.2
metaworld @ git+https://github.com/Farama-Foundation/Metaworld.git@c822f28f582ba1ad49eb5dcf61016566f28003ba
ml-collections==0.1.1
ml-dtypes==0.3.2
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.0.8
mujoco==2.3.7
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
open-clip-torch==2.24.0
opencv-python==4.9.0.80
opt-einsum==3.3.0
optax==0.2.1
orbax-checkpoint==0.5.3
packaging==23.2
pandas==2.2.1
parso==0.8.3
pettingzoo==1.24.3
pexpect==4.9.0
pillow==10.2.0
proglog==0.1.10
prompt-toolkit==3.0.43
protobuf==4.25.3
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
PyOpenGL==3.1.7
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
rich==13.7.1
safetensors==0.4.2
scipy==1.12.0
sentencepiece==0.2.0
six==1.16.0
smmap==5.0.1
soupsieve==2.5
stack-data==0.6.3
sympy==1.12
tensorflow-probability==0.23.0
tensorstore==0.1.54
timm==0.9.16
toolz==0.12.1
torch==2.2.1
torchvision==0.17.1
tqdm==4.66.2
traitlets==5.14.1
triton==2.2.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
zipp==3.17.0